Update package versions and ensure non-null return
This commit updates multiple package references in NugetNinja to their latest versions. This not only includes the version bump in Aiursoft.NugetNinja.PrBot and Aiursoft.NugetNinja.MergeBot itself but also the dependencies, to ensure compatibility and use of the latest features or patches. The nullability issue in GitLabService.cs has been fixed to ensure a non-null DetailedMergeRequest object is always returned.